backend/requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiohttp_autoreload==0.0.1
aiosignal==1.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.1
django-checkdb==1.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-nanoid-field==0.1.4
django-newsingleton==1.1
django-otp==1.5.4
django-svg-image-form-field==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jwcrypto==1.5.6
keyring==25.2.1
loguru==0.7.2
magic-filter==1.0.12
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.3.0
multidict==6.0.5
nanoid==2.0.0
nh3==0.2.17
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pkginfo==1.10.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pypng==0.20220715.0
pytest==8.3.4
pytest-django==4.9.0
qrcode==7.4.2
readme_renderer==43.0
requests==2.32.3
requests-toolbelt==1.0.0
resend==2.4.0
rfc3986==2.0.0
rich==13.7.1
setuptools==70.1.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0
urllib3==2.2.2
user-agents==2.2.0
uvicorn==0.30.1
watchdog==4.0.1
whitenoise==6.7.0
yarl==1.9.4
zipp==3.19.2
